Massive refactor.
