+++ client/src/App.js
@@ -1,10 +1,10 @@
 import React, { useState } from "react"
 import "./App.css"
-import FlightData from "./pages/FlightData.js"
-import FlightPlan from "./pages/FlightPlan.js"
-import Params from "./pages/Params.js"
-import Submissions from "./pages/Submissions.js"
-import AntennaTracker from "./pages/AntennaTracker.js"
+import FlightData from "pages/FlightData"
+import FlightPlan from "pages/FlightPlan"
+import Params from "pages/Params"
+import Submissions from "pages/Submissions"
+import AntennaTracker from "pages/AntennaTracker.js"
 import "bootstrap/dist/css/bootstrap.min.css"
 import { darker } from "./theme/Colors"
 import { BrowserRouter as Router, Switch, Route, Redirect } from "react-router-dom"
