+++ client/src/components/Containers/index.js
@@ -1,0 +1,2 @@
+export { default as Column } from "./Column"
+export { default as Row } from "./Row"
