+++ client/src/components/NavBar.js
@@ -1,6 +1,6 @@
 import React from "react"
-import { Row, Column } from "./Containers"
-import { darker } from "../theme/Colors"
+import { Row } from "components/Containers"
+import { darker } from "theme/Colors"
 import styled from "styled-components"
 
 const NavContainer = styled.div`
