+++ client/src/components/UIElements/index.js
@@ -1,0 +1,7 @@
+export { default as Box } from "./Box"
+export { default as Button } from "./Button"
+export { default as Checkbox } from "./Checkbox"
+export { default as Dropdown } from "./Dropdown"
+export { default as Label } from "./Label"
+export { default as Link } from "./Link"
+export { default as Slider } from "./Slider"
