+++ client/src/pages/Params/Param/Value.js
@@ -1,0 +1,25 @@
+import React from "react"
+
+import { dark } from "theme/Colors"
+
+const Value = ({ hook, editable }) => {
+	const [value, setValue] = hook
+
+	return (
+		<input
+			type="text"
+			name="value"
+			value={value}
+			onChange={e => setValue(e.target.value)}
+			style={{
+				background: dark,
+				border: "none",
+				outline: "none",
+				textAlign: "center",
+			}}
+			readOnly={editable ? false : true}
+		/>
+	)
+}
+
+export default Value
